+++ packages/uxpin-merge-cli/src/program/command/init/getInitCommandSteps.ts
@@ -13,10 +13,16 @@ export interface DefaultFile {
   target:PathLike;
 }
 
+const REQUIRED_DEPENDENCIES:string[] = [
+  'babel-loader',
+  '@babel/preset-env',
+  'prop-types',
+];
+
 const DEFAULT_CONFIG_FILES:DefaultFile[] = [
   { source: 'uxpin.config.js', target: 'uxpin.config.js' },
   { source: 'uxpin.webpack.config.js', target: 'uxpin.webpack.config.js' },
-  { source: 'UXPinWrapper.js', target: 'src/Wrapper/UXPinWrapper.js' },
+  { source: 'UXPinWrapper.js', target: 'src/components/UXPinWrapper/UXPinWrapper.js' },
 ];
 
 const EXAMPLE_COMPONENT:DefaultFile = {
@@ -27,6 +33,7 @@ const EXAMPLE_COMPONENT:DefaultFile = {
 export function getInitCommandSteps(args:InitProgramArgs):Step[] {
   return [
     { exec: copyDefaultFiles(args), shouldRun: true },
+    { exec: checkDependencies(), shouldRun: true },
   ];
 }
 
@@ -35,19 +42,32 @@ function copyDefaultFiles(args:InitProgramArgs):any {
   try {
     // config files
     DEFAULT_CONFIG_FILES.forEach((file) => {
-      const filePath:PathLike = projectRoot + '/' + file.target;
+      const filePath:PathLike = `${projectRoot}/${file.target}`;
       if (!existsSync(filePath)) {
         copySync(resolve(__dirname, `${RESOURCES_PATH}/${file.source}`), filePath);
+        printLine(`✅ Successfully created ${filePath}`, { color: PrintColor.GREEN });
       }
     });
 
     // default component
-    const componentPath:PathLike = projectRoot + '/' + EXAMPLE_COMPONENT.target;
+    const componentPath:PathLike = `${projectRoot}/${EXAMPLE_COMPONENT.target}`;
     if (!existsSync(componentPath)) {
       copySync(resolve(__dirname, `${RESOURCES_PATH}/${EXAMPLE_COMPONENT.source}`), componentPath);
+      printLine(`✅ Successfully created example component in ${componentPath}`, { color: PrintColor.GREEN });
     }
   } catch (error) {
     printLine('🛑 There was an error while copying default config files. Please try again.', { color: PrintColor.RED });
     throw new Error(error.message);
   }
+}
+
+function checkDependencies():any {
+  REQUIRED_DEPENDENCIES.forEach((library) => {
+    try {
+      require.resolve(library);
+      printLine(`✅ ${library} dependency found.`, { color: PrintColor.GREEN });
+    } catch (e) {
+      printLine(`🛑 ${library} dependency is missing. Please install it.`, { color: PrintColor.RED });
+    }
+  });
 }