+++ packages/uxpin-merge-cli/src/resources/Button/presets/0-default.jsx
@@ -1,17 +1,10 @@
 import * as React from "react";
 import Button from "../Button";
 
-const customFunction = () => "HelloWorld";
-
 export default (
   <Button
     uxpId="button"
-    className="button"
-    name="name"
-    label="label"
-    value="value"
-    onClick={customFunction()}
-    disabled="true">
-    Button
+    disabled="false"
+    label="Button">
   </Button>
 );