+++ packages/uxpin-merge-cli/src/resources/uxpin.config.js
@@ -5,10 +5,10 @@ module.exports = {
         name: 'General',
         include: [
           'src/components/Button/Button.js',
-        ]
-      }
+        ],
+      },
     ],
-    wrapper: 'src/Wrapper/UXPinWrapper.js',
+    wrapper: 'src/components/UXPinWrapper/UXPinWrapper.js',
     webpackConfig: 'uxpin.webpack.config.js',
   },
   name: 'Example Design System'