+++ packages/uxpin-merge-cli/CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [2.7.3] - 2021-01-26
 
 ### Added
-- Added init command which add default files required by merge
+- Added init command which add the default files required by merge
 
 ## [2.7.2] - 2020-12-15